+++ modules/src/ics04_channel/packet.rs
@@ -6,6 +6,7 @@ use ibc_proto::ibc::core::channel::v1::Packet as RawPacket;
 
 use crate::ics04_channel::error::Kind;
 use crate::ics24_host::identifier::{ChannelId, PortId};
+use crate::timestamp::Timestamp;
 use crate::Height;
 
 use super::handler::{
@@ -91,7 +92,7 @@ pub struct Packet {
     #[serde(serialize_with = "crate::serializers::ser_hex_upper")]
     pub data: Vec<u8>,
     pub timeout_height: Height,
-    pub timeout_timestamp: u64,
+    pub timeout_timestamp: Timestamp,
 }
 
 impl std::fmt::Debug for Packet {
@@ -131,7 +132,7 @@ impl Default for Packet {
             destination_channel: Default::default(),
             data: vec![],
             timeout_height: Default::default(),
-            timeout_timestamp: 0,
+            timeout_timestamp: Default::default(),
         }
     }
 }
@@ -156,6 +157,9 @@ impl TryFrom<RawPacket> for Packet {
             return Err(Kind::ZeroPacketData.into());
         }
 
+        let timeout_timestamp = Timestamp::from_nanoseconds(raw_pkt.timeout_timestamp)
+            .map_err(|_| Kind::InvalidPacketTimestamp)?;
+
         Ok(Packet {
             sequence: Sequence::from(raw_pkt.sequence),
             source_port: raw_pkt
@@ -176,7 +180,7 @@ impl TryFrom<RawPacket> for Packet {
                 .map_err(|e| Kind::IdentifierError.context(e))?,
             data: raw_pkt.data,
             timeout_height: packet_timeout_height,
-            timeout_timestamp: raw_pkt.timeout_timestamp,
+            timeout_timestamp,
         })
     }
 }
@@ -191,7 +195,7 @@ impl From<Packet> for RawPacket {
             destination_channel: packet.destination_channel.to_string(),
             data: packet.data,
             timeout_height: Some(packet.timeout_height.into()),
-            timeout_timestamp: packet.timeout_timestamp,
+            timeout_timestamp: packet.timeout_timestamp.as_nanoseconds(),
         }
     }
 }