+++ CHANGELOG.md
@@ -6,9 +6,11 @@
 
 - [ibc]
   - Reinstated `ics23` dependency ([#854])
+  - Use proper Timestamp type to track time ([#855])
+
 - [ibc-relayer]
-  - Change the default for client creation to allow governance recovery in case of expiration or misbehaviour. ([#785])
-  
+  - Change the default for client creation to allow governance recovery in case of expiration or misbehaviour ([#785])
+
 ### BUG FIXES
 
 - [ibc-relayer]
@@ -28,6 +30,7 @@
 [#854]: https://github.com/informalsystems/ibc-rs/issues/854
 [#851]: https://github.com/informalsystems/ibc-rs/issues/851
 [#869]: https://github.com/informalsystems/ibc-rs/issues/869
+[#855]: https://github.com/informalsystems/ibc-rs/issues/855
 
 
 ## v0.2.0